Move chart types to shared folder
